mv rpc_server::tests to tests::integration_tests
