+++ crates/rethnet_rpc_server/src/lib.rs
@@ -22,7 +22,7 @@ use rethnet_evm::{
 };
 
 mod hardhat_methods;
-use hardhat_methods::{reset::RpcHardhatNetworkConfig, HardhatMethodInvocation};
+pub use hardhat_methods::{reset::RpcHardhatNetworkConfig, HardhatMethodInvocation};
 
 /// an RPC method with its parameters
 #[derive(Clone, Debug, PartialEq, serde::Deserialize, serde::Serialize)]
@@ -208,7 +208,7 @@ async fn handle_set_storage_at(
     }
 }
 
-pub async fn router(state: StateType) -> Router {
+async fn router(state: StateType) -> Router {
     Router::new()
         .route(
             "/",
@@ -358,415 +358,4 @@ pub async fn run(
     });
 
     Ok(address)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use rethnet_eth::{remote::BlockSpec, Bytes};
-
-    async fn start_server() -> SocketAddr {
-        let mut accounts: hashbrown::HashMap<Address, AccountInfo> = Default::default();
-        accounts.insert(
-            Address::from_low_u64_ne(1),
-            AccountInfo {
-                code: None,
-                code_hash: KECCAK_EMPTY,
-                balance: U256::ZERO,
-                nonce: 0,
-            },
-        );
-
-        run(
-            "127.0.0.1:0".parse::<SocketAddr>().unwrap(),
-            RpcHardhatNetworkConfig { forking: None },
-            accounts,
-        )
-        .await
-        .unwrap()
-    }
-
-    async fn submit_request(
-        address: &SocketAddr,
-        request: &RpcRequest<MethodInvocation>,
-    ) -> String {
-        let url = format!("http://{address}/");
-        let body = serde_json::to_string(&request).expect("should serialize request to JSON");
-        reqwest::Client::new()
-            .post(&url)
-            .header(reqwest::header::CONTENT_TYPE, "application/json")
-            .body(body.clone())
-            .send()
-            .await
-            .expect(&format!("should send to url '{url}' request body '{body}'"))
-            .text()
-            .await
-            .expect(&format!("should get full response text"))
-    }
-
-    #[tokio::test]
-    async fn test_get_balance_nonexistent_account() {
-        let request = RpcRequest {
-            version: jsonrpc::Version::V2_0,
-            id: jsonrpc::Id::Num(0),
-            method: MethodInvocation::Eth(EthMethodInvocation::GetBalance(
-                Address::from_low_u64_ne(2),
-                BlockSpec::Tag(String::from("latest")),
-            )),
-        };
-
-        let expected_response = jsonrpc::Response::<U256> {
-            jsonrpc: request.version,
-            id: request.id.clone(),
-            data: jsonrpc::ResponseData::Error {
-                error: jsonrpc::Error {
-                    code: 0,
-                    message: String::from("No such account"),
-                    data: None,
-                },
-            },
-        };
-
-        let actual_response: jsonrpc::Response<U256> =
-            serde_json::from_str(&submit_request(&start_server().await, &request).await)
-                .expect("should deserialize from JSON");
-
-        assert_eq!(actual_response, expected_response);
-    }
-
-    #[tokio::test]
-    async fn test_get_balance_success() {
-        let request = RpcRequest {
-            version: jsonrpc::Version::V2_0,
-            id: jsonrpc::Id::Num(0),
-            method: MethodInvocation::Eth(EthMethodInvocation::GetBalance(
-                Address::from_low_u64_ne(1),
-                BlockSpec::Tag(String::from("latest")),
-            )),
-        };
-
-        let expected_response = jsonrpc::Response::<U256> {
-            jsonrpc: request.version,
-            id: request.id.clone(),
-            data: jsonrpc::ResponseData::Success { result: U256::ZERO },
-        };
-
-        let actual_response: jsonrpc::Response<U256> =
-            serde_json::from_str(&submit_request(&start_server().await, &request).await)
-                .expect("should deserialize from JSON");
-
-        assert_eq!(actual_response, expected_response);
-    }
-
-    #[tokio::test]
-    async fn test_get_code_success() {
-        let request = RpcRequest {
-            version: jsonrpc::Version::V2_0,
-            id: jsonrpc::Id::Num(0),
-            method: MethodInvocation::Eth(EthMethodInvocation::GetCode(
-                Address::from_low_u64_ne(1),
-                BlockSpec::Tag(String::from("latest")),
-            )),
-        };
-
-        let expected_response = jsonrpc::Response::<ZeroXPrefixedBytes> {
-            jsonrpc: request.version,
-            id: request.id.clone(),
-            data: jsonrpc::ResponseData::Success {
-                result: ZeroXPrefixedBytes::from(Bytes::from_static(b"\0")),
-            },
-        };
-
-        let actual_response: jsonrpc::Response<ZeroXPrefixedBytes> =
-            serde_json::from_str(&submit_request(&start_server().await, &request).await)
-                .expect("should deserialize from JSON");
-
-        assert_eq!(actual_response, expected_response);
-    }
-
-    #[tokio::test]
-    async fn test_get_storage_success() {
-        let request = RpcRequest {
-            version: jsonrpc::Version::V2_0,
-            id: jsonrpc::Id::Num(0),
-            method: MethodInvocation::Eth(EthMethodInvocation::GetStorageAt(
-                Address::from_low_u64_ne(1),
-                U256::ZERO,
-                BlockSpec::Tag(String::from("latest")),
-            )),
-        };
-
-        let expected_response = jsonrpc::Response::<U256> {
-            jsonrpc: request.version,
-            id: request.id.clone(),
-            data: jsonrpc::ResponseData::Success { result: U256::ZERO },
-        };
-
-        let actual_response: jsonrpc::Response<U256> =
-            serde_json::from_str(&submit_request(&start_server().await, &request).await)
-                .expect("should deserialize from JSON");
-
-        assert_eq!(actual_response, expected_response);
-    }
-
-    #[tokio::test]
-    async fn test_get_transaction_count_nonexistent_account() {
-        let request = RpcRequest {
-            version: jsonrpc::Version::V2_0,
-            id: jsonrpc::Id::Num(0),
-            method: MethodInvocation::Eth(EthMethodInvocation::GetTransactionCount(
-                Address::from_low_u64_ne(2),
-                BlockSpec::Tag(String::from("latest")),
-            )),
-        };
-
-        let expected_response = jsonrpc::Response::<U256> {
-            jsonrpc: request.version,
-            id: request.id.clone(),
-            data: jsonrpc::ResponseData::Error {
-                error: jsonrpc::Error {
-                    code: 0,
-                    message: String::from("No such account"),
-                    data: None,
-                },
-            },
-        };
-
-        let actual_response: jsonrpc::Response<U256> =
-            serde_json::from_str(&submit_request(&start_server().await, &request).await)
-                .expect("should deserialize from JSON");
-
-        assert_eq!(actual_response, expected_response);
-    }
-
-    #[tokio::test]
-    async fn test_get_transaction_count_success() {
-        let request = RpcRequest {
-            version: jsonrpc::Version::V2_0,
-            id: jsonrpc::Id::Num(0),
-            method: MethodInvocation::Eth(EthMethodInvocation::GetTransactionCount(
-                Address::from_low_u64_ne(1),
-                BlockSpec::Tag(String::from("latest")),
-            )),
-        };
-
-        let expected_response = jsonrpc::Response::<U256> {
-            jsonrpc: request.version,
-            id: request.id.clone(),
-            data: jsonrpc::ResponseData::Success { result: U256::ZERO },
-        };
-
-        let actual_response: jsonrpc::Response<U256> =
-            serde_json::from_str(&submit_request(&start_server().await, &request).await)
-                .expect("should deserialize from JSON");
-
-        assert_eq!(actual_response, expected_response);
-    }
-
-    #[tokio::test]
-    async fn test_set_balance_success() {
-        let server_address = start_server().await;
-
-        let address = Address::from_low_u64_ne(1);
-        let new_balance = U256::from(100);
-
-        let request = RpcRequest {
-            version: jsonrpc::Version::V2_0,
-            id: jsonrpc::Id::Num(0),
-            method: MethodInvocation::Hardhat(HardhatMethodInvocation::SetBalance(
-                address,
-                new_balance,
-            )),
-        };
-
-        let expected_response = jsonrpc::Response::<()> {
-            jsonrpc: request.version,
-            id: request.id.clone(),
-            data: jsonrpc::ResponseData::Success { result: () },
-        };
-
-        let actual_response: jsonrpc::Response<()> =
-            serde_json::from_str(&submit_request(&server_address, &request).await)
-                .expect("should deserialize from JSON");
-
-        assert_eq!(actual_response, expected_response);
-
-        let request = RpcRequest {
-            version: jsonrpc::Version::V2_0,
-            id: jsonrpc::Id::Num(0),
-            method: MethodInvocation::Eth(EthMethodInvocation::GetBalance(
-                address,
-                BlockSpec::Tag(String::from("latest")),
-            )),
-        };
-
-        let expected_response = jsonrpc::Response::<U256> {
-            jsonrpc: request.version,
-            id: request.id.clone(),
-            data: jsonrpc::ResponseData::Success {
-                result: new_balance,
-            },
-        };
-
-        let actual_response: jsonrpc::Response<U256> =
-            serde_json::from_str(&submit_request(&server_address, &request).await)
-                .expect("should deserialize from JSON");
-
-        assert_eq!(actual_response, expected_response);
-    }
-
-    #[tokio::test]
-    async fn test_set_nonce_success() {
-        let server_address = start_server().await;
-
-        let address = Address::from_low_u64_ne(1);
-        let new_nonce = U256::from(100);
-
-        let request = RpcRequest {
-            version: jsonrpc::Version::V2_0,
-            id: jsonrpc::Id::Num(0),
-            method: MethodInvocation::Hardhat(HardhatMethodInvocation::SetNonce(
-                address, new_nonce,
-            )),
-        };
-
-        let expected_response = jsonrpc::Response::<()> {
-            jsonrpc: request.version,
-            id: request.id.clone(),
-            data: jsonrpc::ResponseData::Success { result: () },
-        };
-
-        let actual_response: jsonrpc::Response<()> =
-            serde_json::from_str(&submit_request(&server_address, &request).await)
-                .expect("should deserialize from JSON");
-
-        assert_eq!(actual_response, expected_response);
-
-        let request = RpcRequest {
-            version: jsonrpc::Version::V2_0,
-            id: jsonrpc::Id::Num(0),
-            method: MethodInvocation::Eth(EthMethodInvocation::GetTransactionCount(
-                address,
-                BlockSpec::Tag(String::from("latest")),
-            )),
-        };
-
-        let expected_response = jsonrpc::Response::<U256> {
-            jsonrpc: request.version,
-            id: request.id.clone(),
-            data: jsonrpc::ResponseData::Success { result: new_nonce },
-        };
-
-        let actual_response: jsonrpc::Response<U256> =
-            serde_json::from_str(&submit_request(&server_address, &request).await)
-                .expect("should deserialize from JSON");
-
-        assert_eq!(actual_response, expected_response);
-    }
-
-    #[tokio::test]
-    async fn test_set_code_success() {
-        let server_address = start_server().await;
-
-        let address = Address::from_low_u64_ne(1);
-        let new_code = ZeroXPrefixedBytes::from(Bytes::from_static(b"deadbeef"));
-
-        let request = RpcRequest {
-            version: jsonrpc::Version::V2_0,
-            id: jsonrpc::Id::Num(0),
-            method: MethodInvocation::Hardhat(HardhatMethodInvocation::SetCode(
-                address,
-                new_code.clone(),
-            )),
-        };
-
-        let expected_response = jsonrpc::Response::<()> {
-            jsonrpc: request.version,
-            id: request.id.clone(),
-            data: jsonrpc::ResponseData::Success { result: () },
-        };
-
-        let actual_response: jsonrpc::Response<()> =
-            serde_json::from_str(&submit_request(&server_address, &request).await)
-                .expect("should deserialize from JSON");
-
-        assert_eq!(actual_response, expected_response);
-
-        let request = RpcRequest {
-            version: jsonrpc::Version::V2_0,
-            id: jsonrpc::Id::Num(0),
-            method: MethodInvocation::Eth(EthMethodInvocation::GetCode(
-                address,
-                BlockSpec::Tag(String::from("latest")),
-            )),
-        };
-
-        let expected_response = jsonrpc::Response::<ZeroXPrefixedBytes> {
-            jsonrpc: request.version,
-            id: request.id.clone(),
-            data: jsonrpc::ResponseData::Success {
-                result: new_code.clone(),
-            },
-        };
-
-        let actual_response: jsonrpc::Response<ZeroXPrefixedBytes> =
-            serde_json::from_str(&submit_request(&server_address, &request).await)
-                .expect("should deserialize from JSON");
-
-        assert_eq!(actual_response, expected_response);
-    }
-
-    #[tokio::test]
-    async fn test_set_storage_at_success() {
-        let server_address = start_server().await;
-
-        let address = Address::from_low_u64_ne(1);
-        let new_storage_value = U256::from(100);
-
-        let request = RpcRequest {
-            version: jsonrpc::Version::V2_0,
-            id: jsonrpc::Id::Num(0),
-            method: MethodInvocation::Hardhat(HardhatMethodInvocation::SetStorageAt(
-                address,
-                U256::ZERO,
-                new_storage_value,
-            )),
-        };
-
-        let expected_response = jsonrpc::Response::<()> {
-            jsonrpc: request.version,
-            id: request.id.clone(),
-            data: jsonrpc::ResponseData::Success { result: () },
-        };
-
-        let actual_response: jsonrpc::Response<()> =
-            serde_json::from_str(&submit_request(&server_address, &request).await)
-                .expect("should deserialize from JSON");
-
-        assert_eq!(actual_response, expected_response);
-
-        let request = RpcRequest {
-            version: jsonrpc::Version::V2_0,
-            id: jsonrpc::Id::Num(0),
-            method: MethodInvocation::Eth(EthMethodInvocation::GetStorageAt(
-                address,
-                U256::ZERO,
-                BlockSpec::Tag(String::from("latest")),
-            )),
-        };
-
-        let expected_response = jsonrpc::Response::<U256> {
-            jsonrpc: request.version,
-            id: request.id.clone(),
-            data: jsonrpc::ResponseData::Success {
-                result: new_storage_value,
-            },
-        };
-
-        let actual_response: jsonrpc::Response<U256> =
-            serde_json::from_str(&submit_request(&server_address, &request).await)
-                .expect("should deserialize from JSON");
-
-        assert_eq!(actual_response, expected_response);
-    }
 }
